requirements.txt
# Core ML
torch>=2.10.0
numpy>=2.4.0
scikit-learn>=1.8.0
pandas>=2.3.0
faiss-cpu>=1.13.0
lightgbm>=4.6.0
tqdm>=4.67.0

# Experiment tracking
mlflow>=3.10.0

# LLM / RAG stack
openai>=2.24.0
langchain-core>=1.2.17
langchain-openai>=1.1.10
langgraph>=1.0.10
langsmith>=0.7.10
python-dotenv>=1.2.0

# Hybrid retrieval
rank-bm25>=0.2.2
cohere>=5.20.0

# LLM evaluation
deepeval>=3.8.8
